Added RX and updated PWM code.

+++ include/datatypes.h
@@ -52,11 +52,11 @@ struct BarometerData {
 
 
 struct RCData {
-    float aileron_pwm;
-    float elevator_pwm;
-    float throttle_pwm;
-    float rudder_pwm;
-    float flightmode_pwm;
+    unsigned int aileron_pwm;
+    unsigned int elevator_pwm;
+    unsigned int throttle_pwm;
+    unsigned int rudder_pwm;
+    unsigned int flightmode_pwm;
     bool healthy; 
 };
